Make sure avatar is always copied to avatars dir unless it already exists

+++ app/models/upload.rb
@@ -30,8 +30,8 @@ class Upload < ApplicationRecord
         extension: "jpeg"
       )
       upload.save!
-      FileUtils.cp(file_path, upload.path)
     end
+    FileUtils.cp(file_path, upload.path) unless File.exists?(upload.path)
     upload
   end
 end
